requirements.txt
numpy~=2.0.2
streamlit~=1.40.1
tensorflow~=2.18.0
pillow~=11.0.0
keras~=3.6.0
scikit-learn~=1.5.2
tqdm~=4.67.0
opencv-python~=4.10.0.84